requirements.txt
# Core Dependencies
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0
scikit-learn>=1.0.0

# Data and Market Connectivity
yfinance>=0.1.70
ccxt>=1.90.0
websocket-client>=1.3.0
requests>=2.28.0

# Technical Analysis
ta-lib>=0.4.0
talib>=0.4.0

# Statistical Analysis
statsmodels>=0.13.0
arch>=5.3.0

# Performance and Optimization
numba>=0.56.0
joblib>=1.1.0

# Database and Storage
sqlalchemy>=1.4.0
redis>=4.3.0
sqlite3

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.10.0

# Web Dashboard
flask>=2.2.0
flask-socketio>=5.3.0
dash>=2.6.0
dash-bootstrap-components>=1.2.0

# Configuration and Logging
pyyaml>=6.0
python-dotenv>=0.19.0
loguru>=0.6.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-mock>=3.8.0

# Development
black>=22.0.0
flake8>=5.0.0
mypy>=0.971
pre-commit>=2.20.0

# Async and Concurrency
asyncio-mqtt>=0.11.0
aiohttp>=3.8.0
aioredis>=2.0.0

# Machine Learning (Optional)
tensorflow>=2.9.0
torch>=1.12.0
lightgbm>=3.3.0
xgboost>=1.6.0

# Risk Management
cvxpy>=1.2.0
quantlib-python>=1.27.0